+++ R/fov.R
@@ -733,9 +733,7 @@ subset.FOV <- function(x, cells = NULL, features = NULL, ...) {
   for (i in Molecules(object = x)) {
     x[[i]] <- subset(x = x[[i]], features = features)
   }
-  if (is.numeric(x = cells)) {
-    cells <- Cells(x = x, boundary = NA)[cells]
-  }
+  cells <- .CellSelection(cells = cells, all = Cells(x = x, boundary = NA))
   for (i in Boundaries(object = x)) {
     x[[i]] <- subset(x = x[[i]], cells = cells)
   }